[transform] move polygons to be in a debug layer in frontend; factor out transform code in frontend
