+++ src/react-renderer/src/Polygon.tsx
@@ -1,15 +1,7 @@
 import * as React from "react";
-import { toHex, toPointListString } from "./Util";
+import { toHex, toPointListString, svgTransformString } from "./Util";
 import draggable from "./Draggable";
 import { IGPIPropsDraggable } from "./types";
-
-const penroseToSVG = (canvasSize: [number, number]) => {
-    const [width, height] = canvasSize;
-    const flipYStr = "matrix(1 0 0 -1 0 0)";
-    const translateStr = "translate(" + width / 2 + ", " + height / 2 + ")";
-    // Flip Y direction, then translate shape to origin mid-canvas
-    return [translateStr, flipYStr].join(" "); 
-};
 
 class Polygon extends React.Component<IGPIPropsDraggable> {
     public render() {
@@ -28,15 +20,7 @@ class Polygon extends React.Component<IGPIPropsDraggable> {
             canvasSize
         );
 
-	const tf = shape.transformation.contents;
-	console.log("transformation", tf);
-	const transformList = [tf.xScale, tf.ySkew, tf.xSkew,
-			       tf.yScale, tf.dx, tf.dy];
-	const penroseTransform = "matrix(" + transformList.join(" ") + ")";
-
-	// Do Penrose transform, then SVG
-	const transformStr = [penroseToSVG(canvasSize), penroseTransform].join(" ");
-	console.log("transformStr", transformStr);
+	const transformStr = svgTransformString(shape.transformation.contents, canvasSize);
 
 	return (
                 <polygon
@@ -50,7 +34,7 @@ class Polygon extends React.Component<IGPIPropsDraggable> {
 		  strokeDasharray={ shape.strokeStyle.contents === "dashed" ? "7, 5" : "" }
 		  strokeWidth={thickness}
 
-		  transform={transformStr}
+	          transform={transformStr}
 
 		  onMouseDown={onClick}
 		>
