+++ src/react-renderer/src/RectangleTransform.tsx
@@ -1,15 +1,7 @@
 import * as React from "react";
-import { toHex, toPointListString } from "./Util";
+import { toHex, svgTransformString } from "./Util";
 import draggable from "./Draggable";
 import { IGPIPropsDraggable } from "./types";
-
-const penroseToSVG = (canvasSize: [number, number]) => {
-    const [width, height] = canvasSize;
-    const flipYStr = "matrix(1 0 0 -1 0 0)";
-    const translateStr = "translate(" + width / 2 + ", " + height / 2 + ")";
-    // Flip Y direction, then translate shape to origin mid-canvas
-    return [translateStr, flipYStr].join(" "); 
-};
 
 class RectangleTransform extends React.Component<IGPIPropsDraggable> {
     public render() {
@@ -17,39 +9,17 @@ class RectangleTransform extends React.Component<IGPIPropsDraggable> {
 	const { canvasSize } = this.props;
 	const { onClick } = this.props;
 
-	// const [x, y] = [shape.x.contents, shape.y.contents];
 	const fillColor = toHex(shape.color.contents);
 	const fillAlpha = shape.color.contents[3];
 	const strokeColor = toHex(shape.strokeColor.contents);
 	const strokeAlpha = shape.strokeColor.contents[3];
 	const thickness = shape.strokeWidth.contents;
 
-	// https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/transform
-	// TODO: check that the transform respects screen space, SVG directionality, and the Style writer's intention
-	// Right now increasing the y will move downward. Does that affect the skew factors?
-
-	// const tf = shape.transform.contents;
-	// This is the *full* translation, incl. default transformation
-	const tf = shape.transformation.contents;
-	console.log("transformation", tf);
-	const transformList = [tf.xScale, tf.ySkew, tf.xSkew,
-			       tf.yScale, tf.dx, tf.dy];
-	const penroseTransform = "matrix(" + transformList.join(" ") + ")";
-
-	// Do Penrose transform, then SVG
-	const transformStr = [penroseToSVG(canvasSize), penroseTransform].join(" ");
-	console.log("transformStr", transformStr);
-
-	// Note: the debug polygon already has the penrose transform applied so we only apply the toSVG transform
-        const ptListString = toPointListString(
-            shape.polygon.contents,
-            canvasSize
-        );
+	const transformStr = svgTransformString(shape.transformation.contents, canvasSize);
 
 	// The default rectangle is an axis-aligned unit square centered at the origin 
 	// Its position, size, angle, etc. is all set by the Penrose transform
 	return (
-		<g>
 		<rect
             x={-0.5}
             y={-0.5}
@@ -69,17 +39,6 @@ class RectangleTransform extends React.Component<IGPIPropsDraggable> {
 		>
 		<title>{shape.name.contents}</title>
 		</rect>
-
-                <polygon
-                  points={ptListString}
-		  fillOpacity="0"
-		  stroke="red"
-
-	    transform={penroseToSVG(canvasSize)}
-
-		  onMouseDown={onClick}
-		/>
-		</g>
 	);
     }
 }
