+++ src/react-renderer/src/layers/PolygonLayer.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
-import {toScreen} from "../Util";
+import { toScreen, toPointListString, penroseToSVG } from "../Util";
 import memoize from "fast-memoize";
-import {ILayerProps} from "../types";
+import { ILayerProps } from "../types";
 
-const toPointListString = memoize(
+const toPointListStringOld = memoize(
   (ptList: any[], canvasSize: [number, number]) =>
     ptList
       .map((coords: [number, number]) => {
@@ -15,13 +15,18 @@ const toPointListString = memoize(
 
 class PolygonLayer extends React.Component<ILayerProps> {
   public render() {
-    const {shapes, canvasSize} = this.props;
-    // TODO: control points
+      const {shapes, canvasSize} = this.props;
+      const polygonColor = "red";
+      const shapesWithPolygons = ["CircleTransform", "RectangleTransform", "Polygon"];
+
     return (
       <g>
         {shapes.map(([name, shape]: [string, any], key: number) => {
+
+	    // Curve
+	    // TODO: show control points of curve
           if (name === "Curve") {
-            const ptListString = toPointListString(
+            const ptListString = toPointListStringOld(
               shape.polyline.contents,
               canvasSize
             );
@@ -29,12 +34,33 @@ class PolygonLayer extends React.Component<ILayerProps> {
               <g key={key}>
                 <polyline
                   points={ptListString}
-                  stroke="black"
+                  stroke={polygonColor}
                   fillOpacity="0"
                 />
               </g>
             );
           }
+
+	    // Note: the debug polygon already has the Penrose transform applied so we only apply the transform to SVG space
+	    // TODO: refactor this when all shapes have "polygon" property
+	    else if (shapesWithPolygons.includes(name)) {
+	       const ptListString = toPointListString(
+		   shape.polygon.contents,
+		   canvasSize
+	       );
+
+		return (
+              <g key={key}>
+                <polygon
+                  points={ptListString}
+		  fillOpacity="0"
+		  stroke={polygonColor}
+		  transform={penroseToSVG(canvasSize)}
+		/>
+	      </g>
+		);
+	    }
+
           return <g key={key}/>;
         })}
       </g>
